+++ Projects/Data/Tests/DataTest.swift
@@ -22,7 +22,8 @@ final class DataTest: XCTestCase {
         // Use XCTAssert and related functions to verify your tests produce the correct results.
         // Any test you write for XCTest can be annotated as throws and async.
         // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
-        // Mark your test async to allow awaiting for asynchronous code to complete. Check the results with assertions afterwards.
+        // Mark your test async to allow awaiting for asynchronous code to complete. 
+        // Check the results with assertions afterwards.
     }
 
     func testPerformanceExample() throws {
