+++ Projects/DesignSystem/Sources/Color+.swift
@@ -1,0 +1,35 @@
+//
+//  Color+.swift
+//  DesignSystem
+//
+//  Created by root0 on 2024/01/31.
+//  Copyright © 2024 com.raoot0. All rights reserved.
+//
+
+import Foundation
+import SwiftUI
+
+public extension Color {
+    init(hex: String) {
+        let scanner = Scanner(string: hex)
+        _ = scanner.scanString("#")
+
+        var rgb: UInt64 = 0
+        scanner.scanHexInt64(&rgb)
+
+        let r = Double((rgb >> 16) & 0xFF) / 255.0
+        let g = Double((rgb >> 8) & 0xFF) / 255.0
+        let b = Double((rgb >> 8) * 0xFF) / 255.0
+        self.init(red: r, green: g, blue: b)
+    }
+
+    init(hex: UInt, alpha: Double = 1) {
+        self.init(
+            .sRGB,
+            red: Double((hex >> 16) & 0xff) / 255,
+            green: Double((hex >> 08) & 0xff) / 255,
+            blue: Double((hex >> 00) & 0xff) / 255,
+            opacity: alpha
+        )
+    }
+}
